Unit Tests: MOTD CRUD First Draft (#13)
Implements the first draft of MOTD CRUD unit tests, revises the OpenAPI spec to address oversights, and patches bugs found from testing. Additionally, this refactors the core of each endpoints logic OUT of the controller it's contained in, into files in the `services` folder.

This tests the following business logic:
- `createMotd`
- `fetchLatestMotd`
- `fetchMotd`
- `removeMotd`
- `updateMotd`

This tests the following endpoints:
- POST `/`
- GET `/`
- GET `/:id`
- DELETE `/:id`
- PATCH `/:id`