Import commits from dotnet repo (#15)
* Git Ignore (#7)

Adds a basic git ignore for dotnet core. Exactly what it says on the tin!

* Template API (#8)

Adds a basic template API to base our application off of.

* Initial DB Models (#9)

Implements an initial model for the MOTD item!

* Initial MongoDB Support (#12)

Implements a basic MOTD Service that connects to MongoDB and handles CRUD. Untested ATM as the boilerplate of the app is not complete.

* MOTD Controller Boilerplate (#15)

Implements basic boilerplate for the MOTD API controller. Must be filled in!

* MOTD default populated (#18)

Implements a function into the MotdService that populates a default MOTD if none already exist in the DB.

* Ports for Frontend Use (#20)

Revises launch settings so that they usethe ports expected by the development frontend for proxying requests.

* Bugfix: `_id` doesn't serialize to `id` (#22)

Fixes a bug where the id field on an MOTD would serialize to `id` rather than `_id`.

* Initial MOTD Controller Implementation (#23)

Implements an initial controller for MOTDs, which works with the frontend. Nearly implemented to spec - boilerplate must be refined to support safer code in the db service.

* Bugfix: Wrong HTTP Method in Update MOTD Route (#26)

Fixes a bug where the Update MOTD route was using POST instead of PATCH.

* Safer MOTD DB Controller (#27)

Revises the implementation of `MessageOfTheDayController` to work a little safer. Functions no longer take in entire MOTD objects and now take in specific necessary fields as needed. The CreatedAt and UpdatedAt fields are also maintained by the MOTD Controller now as well.

* Initial commit

* Updated VSCode settings

* Moved into subfolder

* Reverting changes

* Resets entire branch

* Imports entire dotnet project

* Moved extensions recs into workspaces

* Appled tasks and launch configs to codespace